Clean up map structure
