+++ navigation/lib/util.dart
@@ -1,4 +1,4 @@
-import 'package:map/child_map.dart';
+import 'package:map/generic/generic_map.dart';
 import 'package:navigation/coordinate_model.dart';
 
 LatLng toLatLng(CoordinateModel coordinate) {
